refactor: cleanup explore module fixtures and adapter; closes #439
